refactor(cas): split Tuist auth into oauth, oidc, and registration
The Tuist auth module packed interactive browser login, the Continuous
Integration OpenID Connect exchange, and dynamic client registration
into one oversized impl. Each concern now lives in its own child module;
the core keeps the shared state, endpoint resolution, credential
storage, and the token/login/logout entrypoints. Only methods called
across module boundaries are exposed within the auth module, so the type
stays encapsulated from the rest of the crate. Pure code move: tests are
unchanged.
