+++ crates/once-cas/src/tuist/auth.rs
@@ -1,19 +1,24 @@
-use std::fs;
+//! Tuist authentication: token retrieval plus the login and logout entry
+//! points. The interactive browser OAuth flow lives in [`oauth`], the
+//! Continuous Integration `OpenID Connect` exchange in [`oidc`], and
+//! dynamic client registration with its persistence in [`registration`].
+//! This module owns the shared state, endpoint resolution, and credential
+//! storage those flows build on.
+
+mod oauth;
+mod oidc;
+mod registration;
+
 use std::path::{Path, PathBuf};
 use std::sync::{Arc, Mutex as StdMutex};
-use std::thread;
-use std::time::Duration;
 
-use reqwest::{blocking::Client, StatusCode, Url};
-use schlussel::callback::{build_authorization_url, open_browser};
+use reqwest::Url;
 use schlussel::{
-    build_storage_key, CallbackServer, ClientMetadata, DynamicRegistrationClient, FileStorage,
-    OAuthClient, OAuthConfig, PkcePair, SchlusselError, SessionStorage, Token, TokenRefresher,
+    build_storage_key, CallbackServer, FileStorage, SchlusselError, SessionStorage, Token,
 };
-use serde::{Deserialize, Serialize};
 
 use super::{env_token, join_url, TuistCacheConfig, PROVIDER_NAME};
-use crate::{Digest, Error, Result};
+use crate::{Error, Result};
 
 /// Public OAuth client id for Once's built-in Tuist app. This is not a secret.
 pub const TUIST_APP_OAUTH_CLIENT_ID: &str = "b3298a92-3deb-4f5e-a526-b7ad324979b5";
@@ -24,10 +29,6 @@ const REGISTRATION_PATH: &str = "oauth2/register";
 const TOKEN_PATH: &str = "oauth2/token";
 const AUTHORIZATION_PATH: &str = "oauth2/authorize";
 const OIDC_EXCHANGE_PATH: &str = "api/auth/oidc/token";
-const OIDC_AUDIENCE: &str = "tuist";
-const OIDC_REQUEST_TIMEOUT_SECONDS: u64 = 30;
-const OIDC_EXCHANGE_MAX_RETRIES: usize = 3;
-const OIDC_EXCHANGE_RETRY_BASE_DELAY_MS: u64 = 100;
 const OIDC_TOKEN_REFRESH_WINDOW_SECONDS: u64 = 60;
 const CI_ENVIRONMENT_VARIABLES: &[&str] = &["GITHUB_RUN_ID", "CI", "BUILD_NUMBER"];
 
@@ -132,408 +133,6 @@ impl TuistAuth {
         Ok(existed)
     }
 
-    fn load_valid_token(&self) -> Result<Token> {
-        let key = self.storage_key();
-        let storage = self.storage()?;
-        let client = self.oauth_client(
-            storage,
-            self.resolve_client_id()?,
-            "http://127.0.0.1/callback".to_string(),
-        )?;
-        let refresher = TokenRefresher::new(client)
-            .with_file_locking("once")
-            .map_err(|source| Self::remote_auth_error("configure auth refresh", &source))?;
-        refresher
-            .get_valid_token(&key)
-            .map_err(|source| self.cached_auth_error(source))
-    }
-
-    fn load_stored_access_token(&self) -> Result<Option<String>> {
-        let key = self.storage_key();
-        let Some(token) = self
-            .storage()?
-            .load(&key)
-            .map_err(|source| Self::remote_auth_error("load auth token", &source))?
-        else {
-            return Ok(None);
-        };
-        if token.expires_within(OIDC_TOKEN_REFRESH_WINDOW_SECONDS) {
-            return Ok(None);
-        }
-        Ok(Some(token.access_token))
-    }
-
-    fn login_with_client_id(
-        &self,
-        client_id: &str,
-        open_browser_after_prompt: bool,
-        handler: &mut dyn FnMut(TuistAuthPrompt),
-    ) -> Result<()> {
-        let server = CallbackServer::new(0)
-            .map_err(|source| Self::remote_auth_error("start callback", &source))?;
-        let redirect_uri = server.callback_url();
-        self.authorize(
-            client_id,
-            &redirect_uri,
-            &server,
-            open_browser_after_prompt,
-            handler,
-        )
-    }
-
-    fn authorize(
-        &self,
-        client_id: &str,
-        redirect_uri: &str,
-        server: &CallbackServer,
-        open_browser_after_prompt: bool,
-        handler: &mut dyn FnMut(TuistAuthPrompt),
-    ) -> Result<()> {
-        let pkce = PkcePair::generate();
-        let state = random_state();
-        let authorize_url = build_authorization_url(
-            &self.authorization_endpoint()?,
-            client_id,
-            redirect_uri,
-            None,
-            &state,
-            pkce.challenge(),
-        )
-        .map_err(|source| Self::remote_auth_error("build authorization URL", &source))?;
-
-        handler(TuistAuthPrompt {
-            authorize_url: authorize_url.clone(),
-            redirect_uri: redirect_uri.to_string(),
-            opens_browser: open_browser_after_prompt,
-        });
-        if open_browser_after_prompt {
-            open_browser(&authorize_url)
-                .map_err(|source| Self::remote_auth_error("open browser", &source))?;
-        }
-
-        let callback = server
-            .wait_for_callback(120)
-            .map_err(|source| Self::remote_auth_error("wait for auth callback", &source))?;
-        if callback.state.as_deref() != Some(state.as_str()) {
-            return Err(Self::remote_auth_error(
-                "authorize",
-                &SchlusselError::InvalidState,
-            ));
-        }
-        if callback.error_code.is_some() {
-            return Err(Error::Remote {
-                provider: PROVIDER_NAME,
-                operation: "authorize",
-                message: callback
-                    .error_description
-                    .or(callback.error_code)
-                    .unwrap_or_else(|| "authorization denied".to_string()),
-            });
-        }
-        let code = callback.code.ok_or_else(|| Error::Remote {
-            provider: PROVIDER_NAME,
-            operation: "authorize",
-            message: "missing authorization code".to_string(),
-        })?;
-
-        let key = self.storage_key();
-        let storage = self.storage()?;
-        let client = self.oauth_client(storage, client_id.to_string(), redirect_uri.to_string())?;
-        let token = client
-            .exchange_code(&code, pkce.verifier(), redirect_uri)
-            .map_err(|source| Self::remote_auth_error("exchange authorization code", &source))?;
-        client
-            .save_token(&key, &token)
-            .map_err(|source| Self::remote_auth_error("store auth token", &source))?;
-        Ok(())
-    }
-
-    fn login_with_ci_oidc(&self) -> Result<()> {
-        self.exchange_and_store_ci_oidc_token().map(|_| ())
-    }
-
-    fn exchange_and_store_ci_oidc_token(&self) -> Result<Token> {
-        let _guard = self
-            .ci_exchange_lock
-            .lock()
-            .map_err(|source| Error::Remote {
-                provider: PROVIDER_NAME,
-                operation: "lock Continuous Integration auth token exchange",
-                message: source.to_string(),
-            })?;
-        if let Some(token) = self.cached_ci_token()? {
-            return Ok(token);
-        }
-        if let Ok(token) = self.load_valid_token() {
-            self.remember_ci_token(&token)?;
-            return Ok(token);
-        }
-        let oidc_token = Self::fetch_ci_oidc_token()?;
-        let access_token = self.exchange_oidc_token(&oidc_token)?;
-        self.store_access_token(&access_token)?;
-        self.remember_ci_token(&access_token)?;
-        Ok(access_token)
-    }
-
-    fn fetch_ci_oidc_token() -> Result<String> {
-        if env_is_truthy("GITHUB_ACTIONS") {
-            let request_url = env_token("ACTIONS_ID_TOKEN_REQUEST_URL").ok_or_else(|| {
-                Error::InvalidConfig {
-                    provider: PROVIDER_NAME,
-                    message: "GitHub Actions OpenID Connect token request variables are not set. Set `permissions: id-token: write` in the workflow.".to_string(),
-                }
-            })?;
-            let request_token =
-                env_token("ACTIONS_ID_TOKEN_REQUEST_TOKEN").ok_or_else(|| Error::InvalidConfig {
-                    provider: PROVIDER_NAME,
-                    message: "GitHub Actions OpenID Connect token request variables are not set. Set `permissions: id-token: write` in the workflow.".to_string(),
-                })?;
-            return Self::fetch_github_actions_oidc_token(&request_url, &request_token);
-        }
-
-        if env_is_truthy("CIRCLECI") {
-            return env_token("CIRCLE_OIDC_TOKEN_V2")
-                .or_else(|| env_token("CIRCLE_OIDC_TOKEN"))
-                .ok_or_else(|| Error::InvalidConfig {
-                    provider: PROVIDER_NAME,
-                    message: "CircleCI OpenID Connect token was not found. Enable OpenID Connect for the CircleCI project.".to_string(),
-                });
-        }
-
-        if env_is_truthy("BITRISE_IO") {
-            return env_token("BITRISE_OIDC_ID_TOKEN")
-                .or_else(|| env_token("BITRISE_IDENTITY_TOKEN"))
-                .ok_or_else(|| Error::InvalidConfig {
-                    provider: PROVIDER_NAME,
-                    message: "Bitrise OpenID Connect token was not found. Add the Bitrise identity token step before this step.".to_string(),
-                });
-        }
-
-        Err(Error::InvalidConfig {
-            provider: PROVIDER_NAME,
-            message: "OpenID Connect authentication is not supported in this environment. Supported Continuous Integration providers: GitHub Actions, CircleCI, Bitrise.".to_string(),
-        })
-    }
-
-    fn fetch_github_actions_oidc_token(request_url: &str, request_token: &str) -> Result<String> {
-        let mut url = Url::parse(request_url).map_err(|source| Error::InvalidConfig {
-            provider: PROVIDER_NAME,
-            message: format!(
-                "invalid GitHub Actions OpenID Connect token request URL `{request_url}`: {source}"
-            ),
-        })?;
-        url.query_pairs_mut().append_pair("audience", OIDC_AUDIENCE);
-        let response = Self::oidc_http_client()?
-            .get(url)
-            .bearer_auth(request_token)
-            .send()
-            .map_err(|source| Error::Remote {
-                provider: PROVIDER_NAME,
-                operation: "fetch GitHub Actions OpenID Connect token",
-                message: source.to_string(),
-            })?;
-        let status = response.status();
-        let body = response.text().map_err(|source| Error::Remote {
-            provider: PROVIDER_NAME,
-            operation: "fetch GitHub Actions OpenID Connect token",
-            message: source.to_string(),
-        })?;
-        if !status.is_success() {
-            return Err(Error::Remote {
-                provider: PROVIDER_NAME,
-                operation: "fetch GitHub Actions OpenID Connect token",
-                message: status_body_message(status, &body),
-            });
-        }
-
-        let token_response: OidcTokenResponse =
-            serde_json::from_str(&body).map_err(|source| Error::Remote {
-                provider: PROVIDER_NAME,
-                operation: "decode GitHub Actions OpenID Connect token",
-                message: source.to_string(),
-            })?;
-        let token = token_response.value.trim();
-        if token.is_empty() {
-            return Err(Error::Remote {
-                provider: PROVIDER_NAME,
-                operation: "decode GitHub Actions OpenID Connect token",
-                message: "response did not include a token".to_string(),
-            });
-        }
-        Ok(token.to_string())
-    }
-
-    fn exchange_oidc_token(&self, oidc_token: &str) -> Result<Token> {
-        let client = Self::oidc_http_client()?;
-        let endpoint = self.oidc_exchange_endpoint()?;
-
-        for retry in 0..=OIDC_EXCHANGE_MAX_RETRIES {
-            match Self::exchange_oidc_token_once(&client, endpoint.clone(), oidc_token) {
-                Ok(token) => return Ok(token),
-                Err(error) if retry < OIDC_EXCHANGE_MAX_RETRIES && error.retryable => {
-                    thread::sleep(oidc_exchange_retry_delay(retry));
-                }
-                Err(error) => return Err(error.into_remote_error()),
-            }
-        }
-
-        unreachable!("OpenID Connect exchange retry loop must return")
-    }
-
-    fn exchange_oidc_token_once(
-        client: &Client,
-        endpoint: Url,
-        oidc_token: &str,
-    ) -> std::result::Result<Token, OidcExchangeAttemptError> {
-        let response = client
-            .post(endpoint)
-            .json(&OidcExchangeRequest { token: oidc_token })
-            .send()
-            .map_err(|source| OidcExchangeAttemptError {
-                message: source.to_string(),
-                retryable: true,
-            })?;
-        let status = response.status();
-        let body = response.text().map_err(|source| OidcExchangeAttemptError {
-            message: source.to_string(),
-            retryable: oidc_status_is_retryable(status),
-        })?;
-        if !status.is_success() {
-            return Err(OidcExchangeAttemptError {
-                message: status_body_message(status, &body),
-                retryable: oidc_status_is_retryable(status),
-            });
-        }
-
-        let response: OidcExchangeResponse =
-            serde_json::from_str(&body).map_err(|source| OidcExchangeAttemptError {
-                message: source.to_string(),
-                retryable: false,
-            })?;
-        let access_token = response.access_token.trim();
-        if access_token.is_empty() {
-            return Err(OidcExchangeAttemptError {
-                message: "response did not include an access token".to_string(),
-                retryable: false,
-            });
-        }
-
-        Ok(Token::new(access_token, "Bearer").with_expiration(response.expires_in))
-    }
-
-    fn cached_ci_access_token(&self) -> Result<Option<String>> {
-        Ok(self.cached_ci_token()?.map(|token| token.access_token))
-    }
-
-    fn cached_ci_token(&self) -> Result<Option<Token>> {
-        let mut cached = self
-            .cached_ci_token
-            .lock()
-            .map_err(|source| Error::Remote {
-                provider: PROVIDER_NAME,
-                operation: "load cached Continuous Integration auth token",
-                message: source.to_string(),
-            })?;
-        let Some(token) = cached.as_ref() else {
-            return Ok(None);
-        };
-        if !token.expires_within(OIDC_TOKEN_REFRESH_WINDOW_SECONDS) {
-            return Ok(Some(token.clone()));
-        }
-        *cached = None;
-        Ok(None)
-    }
-
-    fn remember_ci_token(&self, token: &Token) -> Result<()> {
-        let mut cached = self
-            .cached_ci_token
-            .lock()
-            .map_err(|source| Error::Remote {
-                provider: PROVIDER_NAME,
-                operation: "store cached Continuous Integration auth token",
-                message: source.to_string(),
-            })?;
-        *cached = Some(token.clone());
-        Ok(())
-    }
-
-    fn store_access_token(&self, token: &Token) -> Result<()> {
-        self.storage()?
-            .save(&self.storage_key(), token)
-            .map_err(|source| Self::remote_auth_error("store auth token", &source))
-    }
-
-    fn oidc_http_client() -> Result<Client> {
-        Client::builder()
-            .timeout(Duration::from_secs(OIDC_REQUEST_TIMEOUT_SECONDS))
-            .build()
-            .map_err(|source| Error::Remote {
-                provider: PROVIDER_NAME,
-                operation: "configure OpenID Connect client",
-                message: source.to_string(),
-            })
-    }
-
-    fn register_client(&self, redirect_uri: &str) -> Result<RegisteredClient> {
-        let client =
-            DynamicRegistrationClient::new(self.registration_endpoint()?).map_err(|source| {
-                Error::InvalidConfig {
-                    provider: PROVIDER_NAME,
-                    message: format!("invalid Tuist registration endpoint: {source}"),
-                }
-            })?;
-        let response = client
-            .register(&ClientMetadata {
-                client_name: "once".to_string(),
-                redirect_uris: vec![redirect_uri.to_string()],
-                grant_types: vec![
-                    "authorization_code".to_string(),
-                    "refresh_token".to_string(),
-                ],
-                response_types: vec!["code".to_string()],
-                token_endpoint_auth_method: Some("none".to_string()),
-                ..ClientMetadata::default()
-            })
-            .map_err(|source| Self::remote_auth_error("register auth client", &source))?;
-        let client_id = response.client_id.trim();
-        if client_id.is_empty() {
-            return Err(Error::Remote {
-                provider: PROVIDER_NAME,
-                operation: "register auth client",
-                message: "Tuist returned an empty client_id".to_string(),
-            });
-        }
-        Ok(RegisteredClient {
-            client_id: client_id.to_string(),
-        })
-    }
-
-    fn oauth_client(
-        &self,
-        storage: FileStorage,
-        client_id: String,
-        redirect_uri: String,
-    ) -> Result<OAuthClient<FileStorage>> {
-        OAuthClient::new(self.oauth_config(client_id, redirect_uri)?, storage).map_err(|source| {
-            Error::InvalidConfig {
-                provider: PROVIDER_NAME,
-                message: format!("invalid Tuist auth configuration: {source}"),
-            }
-        })
-    }
-
-    fn oauth_config(&self, client_id: String, redirect_uri: String) -> Result<OAuthConfig> {
-        Ok(OAuthConfig {
-            client_id,
-            client_secret: None,
-            authorization_endpoint: self.authorization_endpoint()?,
-            token_endpoint: self.token_endpoint()?,
-            redirect_uri,
-            scope: None,
-            device_authorization_endpoint: None,
-        })
-    }
-
     fn storage(&self) -> Result<FileStorage> {
         FileStorage::with_path(&self.credentials_root).map_err(|source| Error::Remote {
             provider: PROVIDER_NAME,
@@ -550,19 +149,6 @@ impl TuistAuth {
         )
     }
 
-    fn resolve_client_id(&self) -> Result<String> {
-        if let Some(client_id) = self.config.oauth_client_id.clone() {
-            return Ok(client_id);
-        }
-        let registered_client =
-            self.load_registered_client()?
-                .ok_or_else(|| Error::InvalidConfig {
-                    provider: PROVIDER_NAME,
-                    message: self.login_hint(),
-                })?;
-        Ok(registered_client.client_id)
-    }
-
     fn registration_endpoint(&self) -> Result<String> {
         Ok(join_url(&self.config.url, REGISTRATION_PATH)?.to_string())
     }
@@ -577,77 +163,6 @@ impl TuistAuth {
 
     fn oidc_exchange_endpoint(&self) -> Result<Url> {
         join_url(&self.config.url, OIDC_EXCHANGE_PATH)
-    }
-
-    fn load_registered_client(&self) -> Result<Option<RegisteredClient>> {
-        let path = self.registration_path();
-        let bytes = match fs::read(&path) {
-            Ok(bytes) => bytes,
-            Err(source) if source.kind() == std::io::ErrorKind::NotFound => return Ok(None),
-            Err(source) => {
-                return Err(Error::Remote {
-                    provider: PROVIDER_NAME,
-                    operation: "load auth registration",
-                    message: source.to_string(),
-                });
-            }
-        };
-        serde_json::from_slice(&bytes)
-            .map(Some)
-            .map_err(|source| Error::InvalidConfig {
-                provider: PROVIDER_NAME,
-                message: format!(
-                    "corrupt Tuist auth registration at {}: {source}",
-                    path.display()
-                ),
-            })
-    }
-
-    fn save_registered_client(&self, registered_client: &RegisteredClient) -> Result<()> {
-        let path = self.registration_path();
-        let Some(parent) = path.parent() else {
-            return Err(Error::Remote {
-                provider: PROVIDER_NAME,
-                operation: "store auth registration",
-                message: "registration path had no parent directory".to_string(),
-            });
-        };
-        fs::create_dir_all(parent).map_err(|source| Error::Remote {
-            provider: PROVIDER_NAME,
-            operation: "store auth registration",
-            message: source.to_string(),
-        })?;
-        let bytes =
-            serde_json::to_vec_pretty(registered_client).map_err(|source| Error::Remote {
-                provider: PROVIDER_NAME,
-                operation: "store auth registration",
-                message: source.to_string(),
-            })?;
-        fs::write(path, bytes).map_err(|source| Error::Remote {
-            provider: PROVIDER_NAME,
-            operation: "store auth registration",
-            message: source.to_string(),
-        })
-    }
-
-    fn delete_registered_client(&self) -> Result<()> {
-        let path = self.registration_path();
-        match fs::remove_file(path) {
-            Ok(()) => Ok(()),
-            Err(source) if source.kind() == std::io::ErrorKind::NotFound => Ok(()),
-            Err(source) => Err(Error::Remote {
-                provider: PROVIDER_NAME,
-                operation: "delete auth registration",
-                message: source.to_string(),
-            }),
-        }
-    }
-
-    fn registration_path(&self) -> PathBuf {
-        let key = Digest::of_bytes(self.identity().as_bytes()).to_hex();
-        self.credentials_root
-            .join("registrations")
-            .join(format!("{key}.json"))
     }
 
     fn identity(&self) -> String {
@@ -682,79 +197,10 @@ impl TuistAuth {
     }
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-struct RegisteredClient {
-    client_id: String,
-}
-
-#[derive(Debug, Deserialize)]
-struct OidcTokenResponse {
-    value: String,
-}
-
-#[derive(Debug, Serialize)]
-struct OidcExchangeRequest<'a> {
-    token: &'a str,
-}
-
-#[derive(Debug, Deserialize)]
-struct OidcExchangeResponse {
-    access_token: String,
-    #[serde(default)]
-    expires_in: Option<u64>,
-}
-
-#[derive(Debug)]
-struct OidcExchangeAttemptError {
-    message: String,
-    retryable: bool,
-}
-
-impl OidcExchangeAttemptError {
-    fn into_remote_error(self) -> Error {
-        Error::Remote {
-            provider: PROVIDER_NAME,
-            operation: "exchange OpenID Connect token",
-            message: self.message,
-        }
-    }
-}
-
-fn random_state() -> String {
-    PkcePair::generate().verifier()[..22].to_string()
-}
-
 fn is_ci_environment() -> bool {
     CI_ENVIRONMENT_VARIABLES
         .iter()
         .any(|name| std::env::var_os(name).is_some())
-}
-
-fn env_is_truthy(name: &str) -> bool {
-    matches!(
-        std::env::var(name).ok().as_deref(),
-        Some("1" | "true" | "TRUE" | "yes" | "YES")
-    )
-}
-
-fn status_body_message(status: StatusCode, body: &str) -> String {
-    let body = body.trim();
-    if body.is_empty() {
-        format!("HTTP {}", status.as_u16())
-    } else {
-        format!("HTTP {}: {body}", status.as_u16())
-    }
-}
-
-fn oidc_status_is_retryable(status: StatusCode) -> bool {
-    status == StatusCode::REQUEST_TIMEOUT
-        || status == StatusCode::TOO_MANY_REQUESTS
-        || status.is_server_error()
-}
-
-fn oidc_exchange_retry_delay(retry: usize) -> Duration {
-    let multiplier = (0..retry).fold(1_u64, |multiplier, _| multiplier.saturating_mul(2));
-    Duration::from_millis(OIDC_EXCHANGE_RETRY_BASE_DELAY_MS.saturating_mul(multiplier))
 }
 
 fn print_prompt(prompt: &TuistAuthPrompt) {
@@ -771,6 +217,7 @@ fn print_prompt(prompt: &TuistAuthPrompt) {
 
 #[cfg(test)]
 mod tests {
+    use super::registration::RegisteredClient;
     use super::*;
     use std::io::{Read, Write};
     use std::net::TcpListener;
